// src/app/backend/types/models/entity.d.ts
type CreateClient = {
  fullName: string;
  fullSurname: string;
  identified: string;
  clientContact: {
    phoneNumber: string;
    email: string;
    address?: string;
  };
  clientVehicle?: [
    {
      brand: string;
      model: string;
      year: number;
      plates: string;
      engineDisplacement: number;
      description?: string;
    }
  ];
};